https://mosipid.atlassian.net/browse/Merge rem

Former-commit-id: 86f941df67c278262a6eef3f67bfeba0c650c788
Former-commit-id: 83ec79f8dd891e0420bf87ea040c61552ff3d44c

+++ kernel/kernel-masterdata-service/src/test/java/io/mosip/kernel/masterdata/test/service/MasterDataServiceTest.java
@@ -346,7 +346,7 @@ public class MasterDataServiceTest {
 		locationHierarchy = new Location();
 		locationHierarchy.setCode("IND");
 		locationHierarchy.setName("INDIA");
-		locationHierarchy.setHierarchyLevel(0);
+		locationHierarchy.setHierarchyLevel((short)0);
 		locationHierarchy.setHierarchyName("country");
 		locationHierarchy.setParentLocCode(null);
 		locationHierarchy.setLangCode("HIN");
@@ -357,7 +357,7 @@ public class MasterDataServiceTest {
 		locationHierarchy1 = new Location();
 		locationHierarchy1.setCode("KAR");
 		locationHierarchy1.setName("KARNATAKA");
-		locationHierarchy1.setHierarchyLevel(1);
+		locationHierarchy1.setHierarchyLevel((short)1);
 		locationHierarchy1.setHierarchyName(null);
 		locationHierarchy1.setParentLocCode("TEST");
 		locationHierarchy1.setLangCode("KAN");
